route-registration-tool/pyproject.toml
[project]
name = "route-registration-tool"
version = "0.1.0"
description = ""
authors = [
    {name = "Lepton Software",email = "team@leptonmaps.com"}
]
readme = "README.md"
requires-python = ">=3.12,<4.0"
dependencies = [
    "fastapi (>=0.115.8)",
    "uvicorn[standard] (>=0.34.0)",
    "python-dotenv (>=1.0.1)",
    "pydantic (>=2.10.6)",
    "httpx (>=0.28.1)",
    "shapely (>=2.0.7)",
    "aiosqlite (>=0.21.0)",
    "sqlalchemy (>=2.0.37)",
    "greenlet (>=3.0.0)",
    "google-cloud-resource-manager (>=1.14.0)",
    "polyline (>=2.0.2)",
    "pyproj (>=3.7.0)",
    "python-multipart (>=0.0.20)",
    "google-cloud-bigquery (>=3.27.0)",
    "geopy (>=2.4.1)",
    "gdal (>=3.8.4)",
    "google-cloud-firestore (>=2.19.0)",
    "aiohttp (>=3.11.11)",
    "requests (>=2.32.3)",
    "geographiclib (>=2.0)",
    "asyncpg (>=0.30.0)",
    "psycopg[binary] (>=3.2.0)",
    "alembic (>=1.14.0)"
]

[tool.poetry.group.dev.dependencies]
pytest = ">=8.0.0"
pytest-asyncio = ">=0.25.0"

[tool.pytest.ini_options]
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "function"
testpaths = ["tests"]
pythonpath = ["."]

[tool.poetry]
package-mode = false

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
